[utils] move timeit to perf
